requirements.txt
numpy
openai==0.27.7
pandas
pinecone-client
scipy
python-dotenv